update build setup and make UFE headers and library available to plugin code when present
This makes the MayaLiveLinkPlugin build check for the presence of UFE headers in the Maya
(or Maya SDK) installation and makes them and the UFE library available when present.
The plugin code can then use the __has_include preprocessor operator to conditionally
compile in support for UFE.

The UFE headers and libraries are assumed to be installed at the location suggested in the
devkit installation documentation:

<Maya installation>\devkit\ufe

Note that only the UFE version 2 library is made available for linking, which means that UFE
is only supported with Maya 2022 and later which support UFE version 2.

The MayaLiveLinkPlugin itself does not yet make use of the UFE API, but it may prove useful
in the future when working with Maya's USD integration.