add command for querying and editing messaging endpoints
This add a LiveLinkMessagingSettings command that can be used to query and
modify the unicast and static UdpMessaging endpoints. A subsequent change
will add a simple UI on the Python side of the plugin that will make use of this
new command. This offers similar functionality to what's available in the
MotionBuilder LiveLink plugin.