ignore non-DAG Maya nodes in selection when adding to LiveLink
The LiveLinkAddSelectionCommand currently expects to work with Maya DAG nodes
when searching for joints, cameras, lights, or transform nodes to add to LiveLink.
If the selection in Maya includes any non-DAG nodes (for example a blendShape
node or any other DG node type like materials, shaders, etc.), using that node to
reset an MItDag iterator is considered invalid and will cause iteration to start from
the root of the Maya scene. This will result in seemingly arbitrary nodes outside of
the selection being added to LiveLink (often the "|persp" camera, since that is
usually the first DAG node visited).