reset unicast endpoint to default when text field is emptied
The Network Endpoints settings dialog currently requires you to explicitly type out the
default unicast endpoint of "0.0.0.0:0" if you want to go back to the default. Otherwise,
trying to reset it by clearing the text field will result in the plugin complaining that an
empty string is not a valid endpoint.

This fixes that so that emptying the text field returns the unicast endpoint setting to the
default.